gccrs const folding port: continue porting potential_constant_expression_1()
This changeset ports cp_global_trees structure which seems to be used
throughout the cp constexpr.cc code. I am not sure what it's purpose is but it
seems it is used to add and manipulate tree information during the c++ compilation process.
It is possible this is not needed in the Rust code and may be taken out later. Also,
the initiliazation function isn't being called from anywhere yet, we will need to find a
suitable process for it.

Signed-off-by: Faisal Abbas <90.abbasfaisal@gmail.com>